refactor: move components to shared dir
