+++ src/shared/components/layout/footer.component.jsx
