+++ src/shared/components/layout/hero.component.jsx
